fdupes.py - adding statistics: count, size, approx. disk usage.

+++ missing/fdupes.py
@@ -67,14 +67,33 @@ finally:
 print('Done in {:.1f} s. ({} dirs, {} files){}'.format(
     time.time() - walk_start, dirs, len(files), ' '*32))
 
+total_waste = 0
+total_by4kb = 0
+total_files = 0
+group_count = 0
+
+
 for (md5, size), dup_files in itertools.groupby(sorted(files), lambda t: t[:2]):
     group = list(dup_files)
     # print(md5, group)
     if len(group) == 1:
         continue
+
+    # some statistics
+    group_count += 1
+    total_files += len(group)-1
+    total_waste += (len(group)-1) * size
+    total_by4kb += ((len(group)-1) * size + 2**12 - 1) / (2**12)
+
     filenames = sorted((g[2] for g in group), key = suitability_max_len_penalize_spaces)
-    print('{} {}\n{}\n'.format(md5, size, '\n'.join(filenames)))
+    print('{}. {} {}\n{}\n'.format(group_count, md5, size, '\n'.join(filenames)))
 
     if do_delete:
+        if size == 0:
+            print('Not deleting empty files.')
+            continue
         for f in filenames[1:]:
             os.unlink(f)
+
+print('Total waste: {:,} Bytes; {:,} KB (4KB-blocks) in {} extra files.'.format(
+    total_waste, total_by4kb * 4, total_files))
